FR translation
bev have done a complete translation. I push it there and let you proofread it.
I did also checked your proposition to bev onces and merged them.
I strongly suggest you to be on the AltF4 discord to be able to check the translation status (i.e. if others are already on it)