ALTF4-19: french translation (#357)
* 2021-01-01-ALTF4-19: french translation

* FR translation

bev have done a complete translation. I push it there and let you proofread it.
I did also checked your proposition to bev onces and merged them.
I strongly suggest you to be on the AltF4 discord to be able to check the translation status (i.e. if others are already on it)

* Update 2021-01-01-ALTF4-19.md

* Fixed quote

Co-authored-by: Alarig Le Lay <alarig.le-lay@gandi.net>
Co-authored-by: Firerazer <36164772+Firerazer@users.noreply.github.com>
Co-authored-by: Conor <me@conorn.xyz>