Update 2021-01-15-ALTF4-21.md
Most of the fixes are related to an issue with my automatic formatter that includes spaces between "//" and "](", which means that many links were never linked and that others did not lead anywhere.

Changed the translation from "Vagões-Tanque" to "vagões de fluido", this is not an error but the second is used in game.

Changed the translation from "roboports pessoais" to "portos robóticos pessoais", this is not an error but the second is used in game.

Some small fixes related to concordance.

Fixed the "Lua" tag to "lua", with that the snippet now has formating highlight as expected