initial commit
This is kinda a rough outline of how I wanted to go about writing this; the trouble is that there's so much I could say about what I've made that I fear overstaying my welcome, especially since I'm quite over the average size entry already. This is how I wanted to illustrate the sequence of events, and I want to use the introduction as a kind of microchosm of what happened when I started making the module. Let me know if it would be better off without it, though.